pyproject.toml
[project]
name = "ganusa-research-agent"
version = "0.1.0"
description = "Research Flow AI agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "langgraph>=0.0.1",
    "langchain-core>=0.2.0",
    "requests",
    "python-dotenv",
    "pydantic>=2.7.0",
    "tavily-python",
    "boto3",
    "jinja2",
    "aiofiles"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
